refactor(handlers): extract transport.ts handler from intent.ts (Stage B)
handleSAPTransport + sibling-candidate constants. 3672 tests, A1 zero-diff.
